removed the rendering file, rendered useless by new interface file. explination for actions are as follows:
Im attempting to improve my workflow by making everything a function.
this should make it a bit easier on me to render things out and what
not. it also improves the ability to debug on the fly, as i can now see
what functions run and which ones dont within the terminal.

Hopefully some good comes from this wasted day.